+++ lib/transport_hal/esp32/udp_esp32.cpp
@@ -12,7 +12,7 @@ WiFiUDP Udp;
  * @param msg - The message to be sent.
  * @return void
  */
-void sendMessage(int address[4], const char * msg){
+void sendMessage(uint8_t address[4], const char * msg){
     IPAddress address_;
     address_[0] = address[0];address_[1] = address[1];
     address_[2] = address[2];address_[3] = address[3];
