+++ lib/transport_hal/udp_interface.h
@@ -1,10 +1,11 @@
 #ifndef UDP_INTERFACE_H
 #define UDP_INTERFACE_H
 
-
 #define UDP_PORT 12345
+#include <cstdint>
 
-void sendMessage(int address[4], const char * msg);
+
+void sendMessage(uint8_t address[4], const char * msg);
 int receiveMessage(char* buffer, size_t bufferSize);
 void beginTransport();
