+++ lib/wifi_hal/esp32/wifi_esp32.cpp
@@ -8,7 +8,7 @@ int parentDisconnectionCount = 0;
 
 
 void (*parentDisconnectCallback)() = nullptr;
-bool (*isChildRegisteredCallback)(int*) = nullptr;
+bool (*isChildRegisteredCallback)(uint8_t*) = nullptr;
 
 
 /**
@@ -21,7 +21,7 @@ bool (*isChildRegisteredCallback)(int*) = nullptr;
  */
 void onSoftAPModeStationConnectedHandler(WiFiEvent_t event, WiFiEventInfo_t info){
     Serial.println("\n[WIFI_EVENTS] Station connected\n");
-    int lostChildMAC[6];
+    uint8_t lostChildMAC[6];
     lostChildMAC[0] = info.wifi_ap_staconnected.mac[0];lostChildMAC[1] = info.wifi_ap_staconnected.mac[1];
     lostChildMAC[2] = info.wifi_ap_staconnected.mac[2];lostChildMAC[3] = info.wifi_ap_staconnected.mac[3];
     lostChildMAC[4] = info.wifi_ap_staconnected.mac[4];lostChildMAC[5] = info.wifi_ap_staconnected.mac[5];
@@ -43,7 +43,7 @@ void onSoftAPModeStationConnectedHandler(WiFiEvent_t event, WiFiEventInfo_t info
  */
 void onSoftAPModeStationDisconnectedHandler(WiFiEvent_t event, WiFiEventInfo_t info) {
     Serial.println("\n[WIFI_EVENTS] Station disconnected\n");
-    int lostChildMAC[6];
+    uint8_t lostChildMAC[6];
     lostChildMAC[0] = info.wifi_ap_stadisconnected.mac[0];lostChildMAC[1] = info.wifi_ap_stadisconnected.mac[1];
     lostChildMAC[2] = info.wifi_ap_stadisconnected.mac[2];lostChildMAC[3] = info.wifi_ap_stadisconnected.mac[3];
     lostChildMAC[4] = info.wifi_ap_stadisconnected.mac[4];lostChildMAC[5] = info.wifi_ap_stadisconnected.mac[5];
@@ -214,7 +214,7 @@ void startWifiSTA(int* localIP, int* gateway, int* subnet, int* dns){
  *
  * @return void
  */
-void startWifiAP(const char* SSID, const char* Pass, int* localIP, int* gateway, int* subnet){
+void startWifiAP(const char* SSID, const char* Pass, uint8_t * localIP, uint8_t* gateway, uint8_t* subnet){
 
     // Set the Wi-Fi mode to operate as both an Access Point (AP) and Station (STA)
     //WiFi.config(localIP, gateway, subnet, gateway);
@@ -236,7 +236,6 @@ void startWifiAP(const char* SSID, const char* Pass, int* localIP, int* gateway,
     initAuxTables();
 
     Serial.printf("My MAC: %s\n", WiFi.macAddress().c_str());
-
 }
 
 /**
@@ -297,7 +296,6 @@ void connectToAP(const char * SSID, const char * PASS) {
     }
     endTime = getCurrentTime();
 
-
     LOG(NETWORK,DEBUG,"WiFi connect time:%lu\n",endTime-startTime);
 }
 /**
@@ -343,7 +341,7 @@ int numberOfSTAConnected(){
  * @param IP - the array that will store the IP address.
  * @return void
  */
-void getGatewayIP(int *IP){
+void getGatewayIP(uint8_t *IP){
     IPAddress ip = WiFi.gatewayIP();
     IP[0] = ip[0];IP[1] = ip[1];
     IP[2] = ip[2];IP[3] = ip[3];
@@ -356,7 +354,7 @@ void getGatewayIP(int *IP){
  * @param IP - the array that will store the IP address.
  * @return void.
  */
-void getMySTAIP(int* IP){
+void getMySTAIP(uint8_t * IP){
     IPAddress ip =  WiFi.localIP();
     IP[0] = ip[0];IP[1] = ip[1];
     IP[2] = ip[2];IP[3] = ip[3];
@@ -369,13 +367,13 @@ void getMySTAIP(int* IP){
  * @param MAC - the array that will store the MAC address.
  * @return void.
  */
-void getMyMAC(int* MAC){
+void getMyMAC(uint8_t * MAC){
     unsigned int unsignedMAC[6];
     //Converting MAC from hexadecimal to unsigned int
     sscanf(WiFi.macAddress().c_str(),"%x:%x:%x:%x:%x:%x",&unsignedMAC[0],&unsignedMAC[1],&unsignedMAC[2],&unsignedMAC[3],&unsignedMAC[4],&unsignedMAC[5]);
 
     for (int i = 0; i < 6; i++) {
-        MAC[i] = (int)unsignedMAC[i];
+        MAC[i] = (uint8_t)unsignedMAC[i];
     }
 }
 
@@ -386,7 +384,7 @@ void getMyMAC(int* MAC){
  * @param IP - the array that will store the IP address.
  * @return void.
  */
-void getMyAPIP(int *IP){
+void getMyAPIP(uint8_t *IP){
     IPAddress ip =  WiFi.softAPIP();
     IP[0] = ip[0];IP[1] = ip[1];
     IP[2] = ip[2];IP[3] = ip[3];
